MEDIUM: ssl: add an option to set backend SNI from the authority TLV
Add the server config option 'forward-dst-sni', which may be relevant
to backends whose address was set from the destination address of the
client, due to the use of '0.0.0.0' or '*' as the address config.
When the option is enabled, and if the client sends an authority
TLV in the PROXY header, then the SNI for communication with the
backend is set from the value of the TLV.

This makes it possible for clients using the "address forwarding"
feature of haproxy to also communicate via haproxy with backends
using TLS with virtual hosting, by using PROXYv2 and TLV.

If the option is enabled and the authority TLV is present, it
overrides the SNI that was set by the client. If no authority
TLV is sent, then everything proceeds as before.